# src/main/resources/swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"chateau-controller","description":"Chateau Controller"},{"name":"commentaire-controller","description":"Commentaire Controller"},{"name":"note-controller","description":"Note Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/chateau/":{"get":{"tags":["chateau-controller"],"summary":"listAllChateau","operationId":"listAllChateauUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ChateauDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["chateau-controller"],"summary":"newChateau","operationId":"newChateauUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"chateauDto","description":"chateauDto","required":true,"schema":{"$ref":"#/definitions/ChateauWithoutLinkDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ChateauDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/chateau/search":{"get":{"tags":["chateau-controller"],"summary":"searchChateauByName","operationId":"searchChateauByNameUsingGET","produces":["*/*"],"parameters":[{"in":"body","name":"chateauDto","description":"chateauDto","required":true,"schema":{"$ref":"#/definitions/ChateauDto"}}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ChateauDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/chateau/{id}":{"get":{"tags":["chateau-controller"],"summary":"chateauById","operationId":"chateauByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ChateauDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["chateau-controller"],"summary":"deleteChateau","operationId":"deleteChateauUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Chateau"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/commentaire/":{"post":{"tags":["commentaire-controller"],"summary":"newCommentaire","operationId":"newCommentaireUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"commentaireDto","description":"commentaireDto","required":true,"schema":{"$ref":"#/definitions/CommentaireDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Commentaire"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/commentaire/chateau/{id}":{"get":{"tags":["commentaire-controller"],"summary":"commentairesByChateau","operationId":"commentairesByChateauUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/CommentaireDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/commentaire/gestion/{id}":{"get":{"tags":["commentaire-controller"],"summary":"commentaireByUser_Id","operationId":"commentaireByUser_IdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/CommentaireDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/commentaire/{id}":{"put":{"tags":["commentaire-controller"],"summary":"updateCommentaire","operationId":"updateCommentaireUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Commentaire"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingHEAD","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"post":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPUT","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingDELETE","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingOPTIONS","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPATCH","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/note/":{"post":{"tags":["note-controller"],"summary":"addNote","operationId":"addNoteUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"note","description":"note","required":true,"schema":{"$ref":"#/definitions/Note"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Note"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/note/{id}":{"get":{"tags":["note-controller"],"summary":"noteByChateau","operationId":"noteByChateauUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Note"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/":{"get":{"tags":["user-controller"],"summary":"get Users","description":"Get list of users","operationId":"listUserUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["user-controller"],"summary":"newUser","operationId":"newUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userDto","description":"userDto","required":true,"schema":{"$ref":"#/definitions/UserDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/login":{"post":{"tags":["user-controller"],"summary":"login","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userDto","description":"userDto","required":true,"schema":{"$ref":"#/definitions/UserDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/token":{"get":{"tags":["user-controller"],"summary":"userToken","operationId":"userTokenUsingGET","produces":["*/*"],"parameters":[{"name":"Authorization","in":"header","description":"Authorization","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/{id}":{"get":{"tags":["user-controller"],"summary":"userId","operationId":"userIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["user-controller"],"summary":"updateUser","operationId":"updateUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"},{"in":"body","name":"userDto","description":"userDto","required":true,"schema":{"$ref":"#/definitions/UserDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["user-controller"],"summary":"deleteUser","operationId":"deleteUserUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDto"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"Chateau":{"type":"object","properties":{"adresse":{"type":"string"},"codePostal":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"numeroAdresse":{"type":"integer","format":"int32"},"responsable":{"$ref":"#/definitions/User"}},"title":"Chateau"},"ChateauDto":{"type":"object","properties":{"adresse":{"type":"string"},"codePostal":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"numeroAdresse":{"type":"integer","format":"int32"},"responsable":{"$ref":"#/definitions/UserDto"}},"title":"ChateauDto"},"ChateauWithoutLinkDTO":{"type":"object","properties":{"adresse":{"type":"string"},"codePostal":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"numeroAdresse":{"type":"integer","format":"int32"},"responsable":{"type":"integer","format":"int32"}},"title":"ChateauWithoutLinkDTO"},"Commentaire":{"type":"object","properties":{"chateau":{"$ref":"#/definitions/Chateau"},"commentaire":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"user":{"$ref":"#/definitions/User"}},"title":"Commentaire"},"CommentaireDto":{"type":"object","properties":{"chateau":{"$ref":"#/definitions/ChateauDto"},"commentaire":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"user":{"$ref":"#/definitions/UserDto"}},"title":"CommentaireDto"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"Note":{"type":"object","properties":{"chateau":{"$ref":"#/definitions/Chateau"},"id":{"type":"integer","format":"int32"},"note":{"type":"integer","format":"int32"},"user":{"$ref":"#/definitions/User"}},"title":"Note"},"User":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"id":{"type":"integer","format":"int32"},"lastName":{"type":"string"},"password":{"type":"string"},"token":{"type":"string"}},"title":"User"},"UserDto":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"id":{"type":"integer","format":"int32"},"lastName":{"type":"string"},"password":{"type":"string"},"token":{"type":"string"}},"title":"UserDto"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}